New: Commit last key when new key is pressed (multitouch-only)
RevBy: TBD
Details: Before registering the new touchpoint, we try to commit the last
pressed key (if still considered to be active).

Replaced QMap with QVector for the touchpoints, as I somehow managed to leak
memory with QMap. Also added more sanity checks. QMap's operator[] is all nice
and such, but that it inserts key-values pairs if the key couldn't be found can
have surprising side effects (why for example do we store the invalid
touchpoint with id = -1 in this map?).

Let it be known that implementing this felt like a kludge, and many crashes had
to be endured. The touchpoint handling code is full of subtle side effects ...
and there are still some crashes left, especially when multiple touchpoints are
involved, or swiping.