Fix include order of static libs for tests, maliit-keyboard-viewer and maliit-keyboard-plugin
RevBy: TrustMe
MALIIT_KEYBOARD_VIEW_LIB depends on MALIIT_KEYBOARD_LIB. Since we link against
static libs when, for example, building the maliit-keyboard-viewer, the include
order matters. For the linker to know which symbols to use from
MALIIT_KEYBOARD_LIB, it first needs to know all used symbols in
MALIIT_KEYBOARD_VIEW_LIB. Otherwise, symbols that are only provided by but not
used in MALIIT_KEYBOARD_LIB will be stripped by the linker, and the compilation
of maliit-keyboard-viewer will fail.
Same applies to tests that use both libraries and maliit-keyboard-plugin.