Simplify signal connection setup in maliit-keyboard
RevBy: TrustMe
MaliitKeyboard::Setup::connectAll can be used to create all required
connections for a fully functional virtual keyboard. It connects instances of
Glass, LayoutUpdater, Renderer, AbstractTextEditor and Logic::WordEngine.

The functions to connect two specific instances to each other are still
available, for example to be used in tests.