New: Experimental word prediction
RevBy: TrustMe
If presage [P] is installed on the system, maliit-keyboard-plugin can be built
with word prediction support like so:

$ qmake CONFIG+=enable-presage

A project include file makes sure that optional libraries are added after
possible static libraries (required for proper symbol discovery by linker),
config-bottom.pri. It is to be included towards the end of other .pro files.

[P] http://presage.sourceforge.net/