Compose Key, KeyArea from Area, Label and change Layout policy
RevBy: TrustMe
Reuse Area and Label for Key and KeyArea.  New Layout policy allows to set
alignment on Layout element (left, top, right, bottom). All models who
previously reported a rectangle (that was used as geometry information) now
report a size, and optionally, an origin. Keys for example are located in a
KeyArea and thus, have an origin. The Layout itself knows the relationship
between WordRibbon, center panel and extended panel. This allows us to contain
even more layout details with it.