Allow datadirectory for maliit-keyboard to be overriden at runtime
RevBy: NeedsRev
This is important to not try to look up unrelated data files
installed by other builds when running tests.
For some components it may make sense to also allow the data
directories to be set programatically for increased testability,
but this can be done later.
Also uses QString consistently for paths, because they contain
encoded data from the system. We expect UTF8.