Drive rendering test with 'misbehaving' surface factories
RevBy: TrustMe
The MaliitKeyboard::Renderer code kind of assumes that if we can create the
main surface then we can also create the other overlay surfaces for extended
keys and magnifier. This manifests itself in function guards that check the
validity of the main surface, but not the others. As shown with the new test
cases, a slightly 'misbehaving' surface factory can easily trigger more
crashes.