Split out closing the keyboard from framework-controlled hiding
RevBy: TrustMe
When the framework decides to hide the input method
(MAbstractInputMethod::hide) -- for instance because the application's text
entry lost focus -- then we cannot at the same time call
MAbstractInputMethodHost::notifyImInitiatedHiding. This conflict will for
example break focus proxy handling and other advanced focus handling use-cases.
The fix is to introduce a dedicated slot, InputMethod::onKeyboardClosed, which
responds to the swipe gesture to close the keyboard.