Create a new view abstraction for Maliit Keyboard
The new interface allows platform-dependent view implementations. On X11, we
want to handle the view differently as for instance on Wayland. The latter
allows usage of subsurfaces, whereas the former doesn't work too well with
multiple property-driven views:
  X11 expects property changes to take place in a certain order and cannot cope
with rapid changes to XWindows at all (for instance, window placement or window
visibility).
  The Surfaces class handles multiple surfaces and should closely match
Wayland's surface handling (including sub-surfaces). The Canvas class on the
other hand represents one drawable area and targets X11. Therefore, it contains
of a single QQuickView with enough space on top of the main keyboard area to
also draw extended keys and the key magnifier.